cleaning up cyberconnect2, moving all tables to files for easier growth in the future. updating any import references. lastly, adding new models to setup_db.py
